Fix EmbedderTarget order
The comment was saying "luckily MSBuild seems to OR the condition automatically" but unfortunately this is not true.

When running `MSBuild`, the order of the targets is correct:  
✅ GenerateSatelliteAssemblies → EmbedderTarget

When running `dotnet build`, the order of the targets is wrong:  
❌ EmbedderTarget → CoreGenerateSatelliteAssemblies

In order to make sure the EmbedderTarget runs after the satellite assemblies are generated, we use a condition on the `GenerateSatelliteAssembliesForCore` property (which is used to conditionally run both [GenerateSatelliteAssemblies][1] and [CoreGenerateSatelliteAssemblies][2] targets).

[1]: https://github.com/dotnet/msbuild/blob/v16.9.0/src/Tasks/Microsoft.Common.CurrentVersion.targets#L3714
[2]: https://github.com/dotnet/sdk/blob/v5.0.202/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.Sdk.targets#L883